feat: Enhance Bingo Winner data with avatar
Introduces the ability to display the winner's Discord avatar in the admin panel and on the bingo dashboard. This provides a more complete and personalized view of bingo achievements.

Additionally, several state variables in the Admin page have been renamed for better clarity, and the default value for the win date input is now set to the current date.